+++ src/exonviz/__init__.py
@@ -1,5 +1,5 @@
 from .draw import draw_exons, config
 from .exon import Exon, Region
-from .cli import fetch_exons
+from .cli import fetch_exons, get_MANE
 
-__all__ = ["draw_exons", "config", "Exon", "Region", "fetch_exons"]
+__all__ = ["draw_exons", "config", "Exon", "Region", "fetch_exons", "get_MANE"]